RDMs combining functions in new rdm.combine
